requirements.txt
absl-py==0.14.1
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
cachetools==4.2.4
certifi==2020.12.5
chardet @ file:///C:/ci/chardet_1607690654534/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.10.0
Flask @ file:///home/ktietz/src/ci/flask_1611932660458/work
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877484576/work
ipykernel @ file:///C:/ci/ipykernel_1596190155316/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1617121002983/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///C:/ci/jedi_1606914528444/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213356021/work
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1612282606037/work
Markdown==3.3.4
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
np-utils==0.5.12.1
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.3.56
pandas @ file:///C:/ci/pandas_1618365634936/work
parso==0.7.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1617386341487/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==227
PyYAML==5.4.1
pyzmq==20.0.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy @ file:///C:/ci/scipy_1618856128765/work
six @ file:///C:/ci/six_1605187374963/work
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.9.4
tornado @ file:///C:/ci/tornado_1606942392901/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work